Remove warnings about comparison of integer expressions of different signedness (-Wsign-compare)

+++ mex/sources/k_order_welfare/objective_m.cc
@@ -42,7 +42,7 @@ ObjectiveMFile::unpackSparseMatrixAndCopyIntoTwoDMatData(mxArray *sparseMat, Two
   /* Under MATLAB, the following check always holds at equality; under Octave,
      there may be an inequality, because Octave diminishes nzmax if one gives
      zeros in the values vector when calling sparse(). */
-  assert(tdm.nrows() >= mxGetNzmax(sparseMat));
+  assert(tdm.nrows() >= static_cast<int>(mxGetNzmax(sparseMat)));
 
   double *ptr = mxGetPr(sparseMat);
